child hunger messages, added random initialization
Added more child messages to bug the parent at different
stages of being hungry.

Now randomly initializes energy to a value between 5 and 10.

When spawning multiple children if they all have the same random
seed (srand(3)) they will all get the same amount of random energy.
When time() as the seed all the children got the same time after forking.
To fix this the process id (getpid(2)) was used as the seed and
srand was called in the child.